perf: use `vim.api.nvim_set_hl()` for speed
As noted in this thread:

- https://www.reddit.com/r/neovim/comments/sihuq7/psa_now_you_can_set_global_highlight_groups_ie/

we can expect this to be faster. On this machine (mid-2015 15" MacBook
Pro), our `nvim --startuptime vim.log` goes from:

    145.776  000.750  000.146: sourcing /Users/wincent/.config/nvim/pack/bundle/opt/cmp-buffer/after/plugin/cmp_buffer.lua
    151.221  001.641  001.641: sourcing /Users/wincent/.config/nvim/pack/bundle/opt/base16-nvim/colors/base16-bright.lua
    151.785  000.361  000.361: require('indent_blankline.utils')

to:

    140.661  000.604  000.079: sourcing /Users/wincent/.config/nvim/pack/bundle/opt/cmp-buffer/after/plugin/cmp_buffer.lua
    144.820  001.082  001.082: sourcing /Users/wincent/.config/nvim/pack/bundle/opt/base16-nvim/colors/base16-bright.lua
    145.383  000.314  000.314: require('indent_blankline.utils')

Columns are:

    clock   self+sourced   self:  sourced script

So, incredibly scientific results are that we dropped from 1.641ms to
1.082ms with this change. In practice, however, I always find that the
gap between the previous file's clock and the file of interest's clock
tells the full story: if we use that metric, we went from 151.221ms -
145.776ms (ie. 5.445ms) to 144.280ms - 140.661ms (ie. 3.619ms). ie.
there is some unaccounted for time in the "self" metric that probably
corresponds to reading the file and parsing it. Either way, this is a
small win but a gratefully accepted one.